feat: modify delete and move event to models folder
