+++ gateway/app/models/__init__.py
@@ -1,4 +1,5 @@
 """Models"""
 from .user import User, UserType
+from .event import Event
 
-__all__ = ["User", "UserType"]
+__all__ = ["User", "UserType", "Event"]
